requirements.txt
absl-py==2.3.1
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bitsandbytes==0.43.3
bleach==6.2.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
comm @ file:///croot/comm_1709322850197/work
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
debugpy @ file:///croot/debugpy_1690905042057/work
decorator==4.4.2
decord==0.6.0
deepspeed==0.17.5
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
einops==0.6.1
einops-exts==0.0.4
eva-decord==0.6.1
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.116.1
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
ffmpy==0.6.1
filelock==3.19.1
flash_attn==2.7.3
fonttools==4.59.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
gradio==5.42.0
gradio_client==1.11.1
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.1.7
hjson==3.1.0
hope==3.6.6
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
imageio==2.34.0
imageio-ffmpeg==0.4.9
install==1.3.5
iopath==0.1.10
ipykernel @ file:///croot/ipykernel_1705933831282/work
ipython @ file:///croot/ipython_1704833016303/work
ipywidgets==8.1.7
isoduration==20.11.0
jedi @ file:///tmp/build/80754af9/jedi_1644315229345/work
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.1
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client @ file:///croot/jupyter_client_1699455897726/work
jupyter_core @ file:///croot/jupyter_core_1698937308754/work
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kagglehub==0.3.12
kiwisolver==1.4.9
lazy_loader==0.4
Levenshtein==0.27.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mistune==3.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.4.2
ninja==1.13.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.99.9
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
orjson==3.11.2
overrides==7.7.0
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.3.1
pandocfilters==1.5.1
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
peft==0.15.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==11.3.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
portalocker==3.2.0
prettytable==3.16.0
proglog==0.1.12
progressbar==2.5
prometheus_client==0.22.1
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
propcache==0.3.2
protobuf==6.32.0
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments @ file:///croot/pygments_1684279966437/work
pyparsing==3.2.3
pysubs2==1.8.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq @ file:///croot/pyzmq_1705605076900/work
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.27.0
ruff==0.12.9
safehttpx==0.1.6
safetensors==0.6.2
-e git+https://github.com/DAMO-NLP-SG/VideoRefer.git@96fbce3287cd1bc08bf75320ba270bd1a46840a5#egg=SAM_2
scenedetect==0.6.3
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.99
sentry-sdk==2.35.0
shellingham==1.5.4
shortuuid==1.0.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
thrift==0.21.0
tifffile==2025.5.10
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado @ file:///croot/tornado_1696936946304/work
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
typer==0.15.3
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.21.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.0
